+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All PRs to the Wasmer repository must add to this file.
 Blocks of changes will separated by version increments.
 
 ## **[Unreleased]**
+- [#515](https://github.com/wasmerio/wasmer/pull/510) Improved Emscripten dyncalls
 - [#513](https://github.com/wasmerio/wasmer/pull/510) Fix emscripten lseek implementation.
 - [#510](https://github.com/wasmerio/wasmer/pull/510) Simplify construction of floating point constants in LLVM backend. Fix LLVM assertion failure due to definition of %ctx.